% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA4GHclient-package.R
\docType{package}
\name{GA4GHclient-package}
\alias{GA4GHclient-package}
\alias{GA4GHclient}
\title{A Bioconductor package for accessing GA4GH API data server}
\description{
GA4GHclient provides an easy way to access public data servers through Global
Alliance for Genomics and Health (GA4GH) genomics API. It provides low-level
access to GA4GH API and translates response data into Bioconductor-based
class objects.
}
\author{
Welliton Souza, Benilton Carvalho, Cristiane Rocha

Maintainer: Welliton Souza <well309@gmail.com>
}
\keyword{package}
