% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchCallSets.R
\name{searchCallSets}
\alias{searchCallSets}
\title{searchCallSets function}
\usage{
searchCallSets(host, variantSetId, name = NA_character_,
  biosampleId = NA_character_, nrows = Inf, responseSize = NA_integer_)
}
\arguments{
\item{host}{URL of GA4GH API data server.}

\item{variantSetId}{The VariantSet to search.}

\item{name}{Only return call sets with this name (case-sensitive, exact match).}

\item{biosampleId}{Return only call sets generated from the provided BioSample ID.}

\item{nrows}{Number of rows of the data frame returned by this function.
If not defined, the function will return all entries. If the number of
available entries is less than the value of this this parameter, the function
will silently return only the available entries.}

\item{responseSize}{Specifies the number of entries to be returned by the
server until reach the number of rows defined in \code{nrows} parameter or
until get all available entries. If not defined, the server will return the
allowed maximum reponse size. Increasing this the value of this parameter will
reduce the number of requests and reducing the time required. The will not
respect this parameter if the value if larger than its maximum response size.}
}
\value{
\code{\link{DataFrame}} object.
}
\description{
Search for call sets (sample columns of VCF files).
}
\details{
This function requests \code{POST host/callsets/search}.
}
\examples{
host <- "http://1kgenomes.ga4gh.org/"
\dontrun{
datasetId <- searchDatasets(host, nrows = 1)$id
variantSetId <- searchVariantSets(host, datasetId, nrows = 1)$id
searchCallSets(host, variantSetId)
}
}
\seealso{
\code{\link{DataFrame}}, \code{\link{getCallSet}}
}
